moved checkAuthentication to authentication lib
