+++ src/cli/lib/configuration/handle-configuration-setup-sso.mjs
@@ -8,12 +8,11 @@ import { Questioner } from 'question-and-answer'
 import { IdentitystoreClient, GetUserIdCommand } from '@aws-sdk/client-identitystore'
 import { SSOAdminClient } from '@aws-sdk/client-sso-admin'
 
-import { checkAuthentication } from '../check-authentication'
+import { checkAuthentication, getCredentials } from '../../../lib/shared/authentication-lib'
 import { cliSpec } from '../../constants'
 import { DEFAULT_SSO_POLICY_NAME, DEFAULT_SSO_GROUP_NAME } from '../../../lib/shared/constants'
 import { ensureRootOrganization } from './lib/ensure-root-organization'
 import { findIdentityStoreStaged } from '../../../lib/shared/find-identity-store'
-import { getCredentials } from '../../../lib/shared/authentication-lib'
 import { progressLogger } from '../../../lib/shared/progress-logger'
 import { setupSSO } from '../../../lib/actions/setup-sso'
